fix(directory): 폴더 구조 보기 쉽게 변경
