+++ src/app/(dashboard)/dashboard/[id]/page.tsx
