+++ src/app/(dashboard)/dashboard/page.tsx
@@ -1,0 +1,9 @@
+const Page = () => {
+  return (
+    <div className='lg:flex-row lg:w-fit lg:overflow-x-scroll sm:m-[20px] flex m-[12px] flex-col w-full'>
+      <span>여긴 대시보드에요</span>
+    </div>
+  );
+};
+
+export default Page;
