+++ src/app/(dashboard)/mydashboard/page.tsx
@@ -1,0 +1,7 @@
+export default function Dashboard() {
+  return (
+    <section className='flex-1 bg-gray-500 h-screen'>
+      <div>나의 대시보드 Section</div>
+    </section>
+  );
+}
