--- src/app/dashboard/mydashboard/page.tsx
@@ -1,8 +1,0 @@
-const dashboard = () => {
-  return (
-    <section className='flex bg-[#FAFAFA] h-screen'>
-      <div>나의 대시보드 Section</div>
-    </section>
-  );
-};
-export default dashboard;
